// packages/knox-api/src/apis/kcs/v1/kc/device/delete-devices/delete-devices.type.ts
/**
 * Represents the structure of the query parameters for an API call that requires device IDs.
 * @property {string[]} profileIds - IMEIs or serial numbers of the devices to be deleted.
 */
export type KCDeleteDevicesArgs = {
  deviceIds: string[];
};

/**
 * Represents the structure of the query parameters for an API call that requires device IDs.
 * @property {number} failedCount - Count of devices a profile could not be unassigned from
 * @property {number} pendingCount - Count of devices that were successfully selected for profile unassignment
 * @property {number} successCount - Count of devices from which a profile was successfully unassigned
 * @property {number} totalCount - Total number of devices
 */
export type KCDeleteDevicesResponse = {
  failedCount: number;
  pendingCount: number;
  successCount: number;
  totalCount: number;
};